requirements.txt
accelerate
transformers
llama-cpp-python
mistral-common
peft
tqdm
autogen
# eliminé las comillas
autogen-ext[llama-cpp]
# Comenté esto que no sé para qué es
#-r ./llama_cpp/requirements.txt
# NUEVOS:
sentencepiece
torch